feat: default compression added (#7)
* feat: arch and buffer

* feat: edgebreaker

* fix(core/buffer): fixed a bug in 'reader'.

* feat(core/buffer): msb/lsb made configurable and edgebreaker is fixed accordingly

* feat: edgebreaker added

* feat: sequential connectivity compression added

* feat: NdVector made efficient for smaller dimension

* refactor(core): Vector implementation and its unpacking logic changed

* feat: prediction transforms for attributes are added.

* feat: attribute compression added

* fix: merging issues fixed

* feat: added general compression technique 1: rans

* feat: all encoders got integrated

* feat: default compression debugged

+++ src/shared/connectivity/edgebreaker/mod.rs
@@ -14,10 +14,10 @@ pub(crate) fn orientation_of_next_face(prev_face: [usize;3], prev_orientation: b
     }
 }
 
-pub(crate) const SYMBOL_ENCODING_CONFIG_SLOT: usize = 4;
-pub(crate) const NUM_CONNECTED_COMPONENTS_SLOT: usize = 8;
-pub(crate) const NUM_FACES_SLOT: usize = 32;
-pub(crate) const HOLE_SLOT_SIZE: usize = 2;
-pub(crate) const HANDLE_SLOT_SIZE: usize = 2;
-pub(crate) const NUM_VERTICES_IN_HOLE_SLOTS: [usize;4] = [8,12,16,20];
-pub(crate) const HANDLE_METADATA_SLOTS: [usize;4] = [8,12,16,20];
+pub(crate) const SYMBOL_ENCODING_CONFIG_SLOT: u8 = 4;
+pub(crate) const NUM_CONNECTED_COMPONENTS_SLOT: u8 = 8;
+pub(crate) const NUM_FACES_SLOT: u8 = 32;
+pub(crate) const HOLE_SLOT_SIZE: u8 = 2;
+pub(crate) const HANDLE_SLOT_SIZE: u8 = 2;
+pub(crate) const NUM_VERTICES_IN_HOLE_SLOTS: [u8;4] = [8,12,16,20];
+pub(crate) const HANDLE_METADATA_SLOTS: [u8;4] = [8,12,16,20];
